проектная работа 11

+++ src/utils/Api.js
@@ -14,7 +14,6 @@ class Api {
     return Promise.reject(`Ошибка: ${res.status}`);
   }
 
-  //1. Загрузка информации о пользователе с сервера - ГОТОВО
   getUserInfo() {
     return fetch(`${this._link}/users/me`, {
       headers: this._headers,
@@ -22,20 +21,18 @@ class Api {
       .then(res => this._checkResponse(res))
   }
 
-  //3. Редактирование профиля -ГОТОВО
   sendEditedUserData(data) {
     return fetch(`${this._link}/users/me`, {
       method: 'PATCH',
       headers: this._headers,
       body: JSON.stringify({
         name: data.name,
-        about: data.info
+        about: data.about
       })
     })
       .then(res => this._checkResponse(res))
   }
 
-  //9. Обновление аватара пользователя
   updateAvatar({ avatar }) {
     return fetch(`${this._link}/users/me/avatar`, {
       method: 'PATCH',
@@ -45,14 +42,13 @@ class Api {
       .then(res => this._checkResponse(res))
   }
 
-  //2. Загрузка карточек с сервера - ГОТОВО
   getInitialCards() {
     return fetch(`${this._link}/cards`, {
       headers: this._headers,
     })
       .then(res => this._checkResponse(res))
   }
-  //4. Добавление новой карточки - ГОТОВО
+
   addCard(card) {
     return fetch(`${this._link}/cards`, {
       method: 'POST',
@@ -64,30 +60,31 @@ class Api {
     })
       .then(res => this._checkResponse(res))
   }
-  //5. Отображение количества лайков карточки
+
   getLikes() {
     return fetch(`${this._link}/cards`, {
       headers: this._headers,
     })
       .then(res => this._checkResponse(res))
   }
-  // Поставить лайк -ГОТОВО
-  putLike(cardId) {
-    return fetch(`${this._link}/cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    })
-      .then(res => this._checkResponse(res))
+
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._link}/cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      })
+        .then(res => this._checkResponse(res))
+    }
+    else {
+      return fetch(`${this._link}/cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      })
+        .then(res => this._checkResponse(res))
+    }
   }
-  // Удалить лайк карточки -ГОТОВО
-  deleteLike(cardId) {
-    return fetch(`${this._link}/cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    })
-      .then(res => this._checkResponse(res))
-  }
-  //6. Попап удаления карточки
+
   deleteCard(cardId) {
     return fetch(`${this._link}/cards/${cardId}`, {
       method: 'DELETE',
